refactor(backend): extract loadStoredPricingOptions to shared module
The loadStoredPricingOptions function was duplicated in both the user.ts and
journey.ts resolvers with nearly identical logic. Extract it into a new
backend/src/lib/pricingOptions.ts module to eliminate code drift risk
when adding new pricing fields in the future.
