+++ backend/src/resolvers/journey.ts
@@ -125,39 +125,7 @@ function extractPricingSnapshot(options?: InputMaybe<JourneySearchOptions>): {
   return snapshot;
 }
 
-/**
- * Reconstructs PricingOptions from a stored Journey entity's snapshotted fields.
- */
-function loadStoredPricingOptions(dbJourney: {
-  firstClass?: boolean;
-  loyaltyCards?: string;
-  ageGroup?: string;
-  deutschlandTicketDiscount?: boolean;
-}): PricingOptions | undefined {
-  const opts: PricingOptions = {};
-
-  if (dbJourney.firstClass !== undefined) {
-    opts.firstClass = dbJourney.firstClass;
-  }
-
-  if (dbJourney.loyaltyCards) {
-    try {
-      opts.loyaltyCards = JSON.parse(dbJourney.loyaltyCards) as LoyaltyCardData[];
-    } catch {
-      Logger.warn('Failed to parse stored loyaltyCards JSON', { raw: dbJourney.loyaltyCards });
-    }
-  }
-
-  if (dbJourney.ageGroup) {
-    opts.ageGroup = dbJourney.ageGroup;
-  }
-
-  if (dbJourney.deutschlandTicketDiscount !== undefined) {
-    opts.deutschlandTicketDiscount = dbJourney.deutschlandTicketDiscount;
-  }
-
-  return Object.keys(opts).length > 0 ? opts : undefined;
-}
+import { loadStoredPricingOptions } from '../lib/pricingOptions';
 
 /**
  * Resolves the 'journeys' query to fetch available journeys from the Hafas API.
