+++ backend/src/resolvers/user.ts
@@ -25,8 +25,8 @@ import { getMeans } from './journey';
 import { NOTIFICATION_TYPES } from './notificationTypes';
 import { ConditionalCheckFailedException } from '@aws-sdk/client-dynamodb';
 import { LoyaltyCardData, validateLoyaltyCard, VALID_AGE_GROUPS } from '../lib/loyaltyCards';
-import { PricingOptions } from '../managers/DbHafasManager';
 import { LoyaltyCard, LoyaltyCardType } from '../schema/generated/typeDefs.generated';
+import { loadStoredPricingOptions } from '../lib/pricingOptions';
 
 dotenv.config();
 
@@ -52,40 +52,6 @@ function parseLoyaltyCards(raw?: string): LoyaltyCard[] | undefined {
     Logger.warn('Failed to parse stored loyaltyCards JSON', { raw });
     return undefined;
   }
-}
-
-/**
- * Loads stored pricing options from a journey entity's snapshotted fields.
- */
-function loadStoredPricingOptions(dbJourney: {
-  firstClass?: boolean;
-  loyaltyCards?: string;
-  ageGroup?: string;
-  deutschlandTicketDiscount?: boolean;
-}): PricingOptions | undefined {
-  const opts: PricingOptions = {};
-
-  if (dbJourney.firstClass !== undefined) {
-    opts.firstClass = dbJourney.firstClass;
-  }
-
-  if (dbJourney.loyaltyCards) {
-    try {
-      opts.loyaltyCards = JSON.parse(dbJourney.loyaltyCards) as LoyaltyCardData[];
-    } catch {
-      Logger.warn('Failed to parse stored loyaltyCards for pricing', { raw: dbJourney.loyaltyCards });
-    }
-  }
-
-  if (dbJourney.ageGroup) {
-    opts.ageGroup = dbJourney.ageGroup;
-  }
-
-  if (dbJourney.deutschlandTicketDiscount !== undefined) {
-    opts.deutschlandTicketDiscount = dbJourney.deutschlandTicketDiscount;
-  }
-
-  return Object.keys(opts).length > 0 ? opts : undefined;
 }
 
 export const userResolvers: UserResolvers = {
